A first functional queued catalog compiles
This is heinously hackish, but it works.  The client
queues a Catalog::Request object, a compiler instance
(started with the new 'puppet compiler' application)
pulls that request off a queue, compiles the catalog,
queues it for the client, and then the client pulls
that catalog off of the queue.

Issues with this commit:

* It's a big conglomeration of hack, just to prove the idea

* It adds some unrelated stuff

* It completely breaks our existing Queue class

* There aren't any tests for any of the new work

But hey, it works, and it's a decent show at how the
models intersect and fight - queueing vs. indirector.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>