Prototype server-side async compiles
This is the last piece to the coin - server-side asynchronous
compiles, where the client still speaks HTTP to the server, but
internally the server passes compile requests off to a message
bus, and a pool of compiler processes subscribe to that bus,
compile the catalogs, and queue them for sending back to the server.

Caveats:

* As with previous versions, no tests, incomplete, etc.

* No error-handling here - if there's a problem, the client will
  never know about it (nor will the master that the client is talking
  to)

* A single pair (request/response) of queues is used for all processing
  right now, but that will need to be changed to a per-client queue, or
  there will need to be conflict checking in some way

* This doesn't yet support the flexibility to have each compiled catalog
  sent along to a database or some such

* This is a completely different code base from the other two attempts,
  and as such many of the things we learned there are missing here

That being said, it's a much better run than the first one - the client
doesn't need any modification at all, but the server can roughly scale
forever.

Even better, if you wanted to implement this today, you could without a
huge amount of effort.  That is, most of the work is in the new compiler
application and modifications to the existing queue terminus, and you could
just add a new terminus rather than modifying the existing one.  You're
not all the way there, but you're close.

Note that this doesn't work for anything other than the compiler - no facts,
reports, etc.

The next step is to unify this series and the previous one, so the
indirector is using request/response internally, and then support generic
connection between the indirector and a bus, rather than catalog-specific
connection.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>